Batch Update
Changed most of the styling to keep in tune with the Vex Website I made.

Changed how the server is started / stopped within the app.  Also added
server starts / stops when the application enters the foreground /
background.

Also as for views, I tried to make the interfaces work on the 3.5 inch
and 4 inch screen.  Still some issues but the majority have been worked
out.